Consolidate tests, add test for generate_1q_gate_statements
